refactor: getResponseBody를 ResponseFactory로 넘김
